+++ lib/common/page_route_observer_instance.dart
@@ -1,0 +1,3 @@
+import 'listenable_page_observer.dart';
+
+final pageRouteObserverInstance = ListenablePageObserver();